Change price type from Int to BigDecimal to handle big decimals

+++ app/src/main/java/com/drosztmer/expensetracker/data/LocalDataSource.kt
@@ -3,6 +3,7 @@ package com.drosztmer.expensetracker.data
 import com.drosztmer.expensetracker.data.database.ExpenseDao
 import com.drosztmer.expensetracker.data.model.Expense
 import kotlinx.coroutines.flow.Flow
+import java.math.BigDecimal
 import javax.inject.Inject
 
 // LocalDataSource for organizing functions on the local data source (there would be a RemoteDataSource class for network calls)
@@ -30,7 +31,7 @@ class LocalDataSource @Inject constructor(
         expenseDao.deleteAllExpenses()
     }
 
-    fun getSum(): Flow<Int> {
+    fun getSum(): Flow<BigDecimal> {
         return expenseDao.getSum()
     }
 }
